Got network communication working.
Two instances can be launched and both will mine. As one finds a block, it broadcasts it to the peers, which have to add the block.

You can launch the first client with python client.py --networked --port=10000 --mine and the second with python client.py --networked --port=10001 --peers=10000 --mine.